chore: rename and move custom_marker_images.dart
